requirements.txt
# Core AI libraries
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Video/Audio processing
yt-dlp>=2023.0.0
moviepy>=1.0.0
ffmpeg-python>=0.2.0

# Audio separation
demucs>=4.0.0

# Speech recognition
funasr>=1.0.0

# Translation
googletrans==4.0.0rc1
openai>=1.0.0
requests>=2.31.0

# Text-to-speech
edge-tts>=6.1.0

# Web UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydub>=0.25.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
tqdm>=4.65.0
logging>=0.5.1.2

# Optional for Ollama
ollama>=0.1.0